refactor: Refactor PalindromeRec class to use String instead of int array
The PalindromeRec class has been refactored to use a String parameter instead of an int array. This simplifies the code and improves readability. The `fun` method now compares characters of the string instead of array elements. The `reveStr` method has also been added to reverse the input string. The main method now takes user input for a string and checks if it is a palindrome using the `fun` method. The output is "yes" if it is a palindrome and "no" otherwise.